Move ensure_multiindex helper to test utilities
