+++ unitest/utils.py
@@ -9,13 +9,10 @@ generally applicable.
 
 from __future__ import annotations
 
-from ast import literal_eval
 from contextlib import contextmanager
-import pandas as pd
 from sklearn.pipeline import FeatureUnion
 
 __all__ = [
-    "ensure_multiindex",
     "patched_feature_union",
 ]
 
@@ -40,36 +37,4 @@ def patched_feature_union():
         yield
     finally:
         FeatureUnion._hstack = original_hstack
-
-
-def ensure_multiindex(df: pd.DataFrame) -> pd.DataFrame:
-    """Return ``df`` with a simple two-level column :class:`~pandas.MultiIndex`.
-
-    The regression tests and tutorials both need feature tables whose columns
-    behave like the objects that :mod:`mne_features` ordinarily emits.  Some
-    extraction paths flatten the MultiIndex into plain strings (for example
-    when results are saved to disk and reloaded), so this helper restores the
-    structured column layout in a way that is safe for any caller that expects
-    the canonical two-level shape.  Keeping the utility here ensures the
-    behaviour is shared consistently across tests and user-facing examples.
-    """
-
-    if isinstance(df.columns, pd.MultiIndex):
-        return df
-
-    def _normalise(column):
-        if isinstance(column, tuple):
-            return column
-        if isinstance(column, str):
-            try:
-                parsed = literal_eval(column)
-            except (ValueError, SyntaxError):
-                parsed = None
-            if isinstance(parsed, tuple):
-                return parsed
-        return (column, "")
-
-    result = df.copy()
-    result.columns = pd.MultiIndex.from_tuples([_normalise(col) for col in df.columns])
-    return result
